requirements.txt
# Core Dependencies
PyQt6==6.6.1
PyQt6-WebEngine==6.6.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# API Clients
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0
requests==2.31.0

# AI Services
openai==1.6.1
google-generativeai==0.3.2
anthropic==0.8.1

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Visualization
matplotlib==3.8.2
plotly==5.18.0

# Notifications
python-telegram-bot==20.7
slack-sdk==3.26.2

# Security
cryptography==41.0.7
PyJWT==2.8.0
pyotp==2.9.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
python-dateutil==2.8.2
pytz==2023.3.post1

# Export
reportlab==4.0.8
openpyxl==3.1.2
Pillow==10.1.0

# Caching
redis==5.0.1

# Logging
loguru==0.7.2

# Development
pytest==7.4.3
pytest-qt==4.3.1
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0
